clean the code :D

+++ src/utils/utils.cpp
@@ -16,22 +16,21 @@
 bn::Ec1 Utils::compute_digest_pub(std::set<int> intersection, const bn::Ec1 g1, PublicKey *pk){
     using namespace NTL;
     using namespace bn;
-    std::vector<double> array(intersection.begin(), intersection.end());
+    std::vector<int> array(intersection.begin(), intersection.end());
     Ec1 digest = g1*0;
-    if(array.size()==0)
+    if(array.size() == 0)
         return digest;
 
-    ZZ_pX f,poly;
-    poly=ZZ_pX(INIT_MONO,array.size());
+    ZZ_pX f, poly;
+    poly=ZZ_pX(INIT_MONO, array.size());
     vec_ZZ_p c;
     c.SetLength(array.size());
-    for(int i=0;i<array.size();i++)
+    for(int i = 0 ; i < array.size(); i++)
         c[i] = conv<ZZ_p>(-array[i]);
 
-    BuildFromRoots(poly,c);
+    BuildFromRoots(poly, c);
 
-
-    for(int i=0;i<array.size()+1;i++){
+    for(int i = 0; i < array.size() + 1; i++){
         const mie::Vuint temp(zToString(poly[i]));
         digest = digest + pk->pubs_g1[i] * temp;
     }
